+++ server/src/services/technology.service.js
@@ -1,4 +1,5 @@
 import { Technology } from "../models/technology.model.js";
+import { deleteFromCloudinary } from "../utils/cloudinary.js";
 
 export const createTechnology = async (data) => {
   return await Technology.create(data);
@@ -17,5 +18,14 @@ export const updateTechnology = async (id, data) => {
 };
 
 export const deleteTechnology = async (id) => {
+  const tech = await Technology.findById(id);
+
+  if (!tech) return null;
+
+  // Delete icon from Cloudinary if it exists
+  if (tech.icon?.public_id) {
+    await deleteFromCloudinary(tech.icon.public_id);
+  }
+
   return await Technology.findByIdAndDelete(id);
 };
